x86_64 build fixes. Patch by Dan Winship
	* common/gkr-crypto.c:
	* pk/gkr-pk-object-manager.c:
	* pk/gkr-pk-privkey.c:
	* pk/gkr-pk-pubkey.c:
	* pk/gkr-pk-util.c:
	* pk/gkr-pk-util.h:
	* pkcs11/gkr-pkcs11-daemon-session.c:
	* pkix/gkr-pkix-cert.c:
	* pkix/gkr-pkix-der.c: x86_64 build fixes. Patch by Dan Winship

svn path=/trunk/; revision=940

+++ common/gkr-crypto.c
@@ -51,7 +51,8 @@ static int
 no_mem_handler (gpointer unused, size_t sz, unsigned int unknown)
 {
 	/* TODO: Figure out additional arguments */
-	g_error ("couldn't allocate %d bytes of memory", sz);
+	g_error ("couldn't allocate %lu bytes of memory", 
+	         (unsigned long int)sz);
 	return 0;
 }
